fix: refactor json module and add prex
