+++ pollux/core/src/main/scala/org/hyperledger/identus/pollux/core/model/CredentialOfferAttachment.scala
@@ -2,7 +2,8 @@ package org.hyperledger.identus.pollux.core.model
 
 import io.circe.*
 import io.circe.generic.semiauto.*
-import org.hyperledger.identus.pollux.core.model.presentation.{Options, PresentationDefinition}
+import org.hyperledger.identus.pollux.core.model.presentation.Options
+import org.hyperledger.identus.pollux.prex.PresentationDefinition
 
 final case class CredentialOfferAttachment(options: Options, presentation_definition: PresentationDefinition)
