+++ pollux/core/src/main/scala/org/hyperledger/identus/pollux/core/model/schema/CredentialSchema.scala
@@ -8,8 +8,8 @@ import org.hyperledger.identus.pollux.core.model.schema.`type`.{
   CredentialSchemaType
 }
 import org.hyperledger.identus.pollux.core.model.schema.`type`.anoncred.AnoncredSchemaSerDesV1
-import org.hyperledger.identus.pollux.core.model.schema.validator.{JsonSchemaValidator, JsonSchemaValidatorImpl}
 import org.hyperledger.identus.pollux.core.service.URIDereferencer
+import org.hyperledger.identus.shared.json.{JsonSchemaValidator, JsonSchemaValidatorImpl}
 import zio.*
 import zio.json.*
 import zio.json.ast.Json
