+++ pollux/core/src/main/scala/org/hyperledger/identus/pollux/core/model/schema/type/AnoncredSchemaType.scala
@@ -3,13 +3,8 @@ package org.hyperledger.identus.pollux.core.model.schema.`type`
 import com.networknt.schema.*
 import org.hyperledger.identus.pollux.core.model.schema.`type`.anoncred.AnoncredSchemaSerDesV1
 import org.hyperledger.identus.pollux.core.model.schema.`type`.anoncred.AnoncredSchemaSerDesV1.*
-import org.hyperledger.identus.pollux.core.model.schema.validator.{
-  JsonSchemaError,
-  JsonSchemaUtils,
-  JsonSchemaValidatorImpl,
-  SchemaSerDes
-}
 import org.hyperledger.identus.pollux.core.model.schema.Schema
+import org.hyperledger.identus.shared.json.{JsonSchemaError, JsonSchemaUtils, JsonSchemaValidatorImpl, SchemaSerDes}
 import zio.*
 import zio.json.*
