--- pollux/core/src/main/scala/org/hyperledger/identus/pollux/core/model/schema/validator/JsonSchemaValidator.scala
@@ -1,12 +1,0 @@
-package org.hyperledger.identus.pollux.core.model.schema.validator
-
-import com.fasterxml.jackson.databind.JsonNode
-import zio.*
-
-trait JsonSchemaValidator {
-  def validate(claims: String): IO[JsonSchemaError, Unit]
-
-  def validate(claimsJsonNode: JsonNode): IO[JsonSchemaError, Unit] = {
-    validate(claimsJsonNode.toString)
-  }
-}
