+++ pollux/core/src/test/scala/org/hyperledger/identus/pollux/core/model/presentation/PresentationAttachmentSpec.scala
@@ -4,6 +4,7 @@ import io.circe.parser.*
 import io.circe.syntax.*
 import io.circe.Json
 import munit.*
+import org.hyperledger.identus.pollux.prex.*
 
 class PresentationAttachmentSpec extends ZSuite {
