+++ pollux/prex/src/test/resources/pd/filter_by_cred_type.json
@@ -1,0 +1,24 @@
+{
+  "presentation_definition": {
+    "id": "first simple example",
+    "input_descriptors": [
+      {
+        "id": "A specific type of VC",
+        "name": "A specific type of VC",
+        "purpose": "We want a VC of this type",
+        "constraints": {
+          "fields": [
+            {
+              "path": [
+                "$.type"
+              ],
+              "filter": {
+                "type": "array"
+              }
+            }
+          ]
+        }
+      }
+    ]
+  }
+}
