+++ pollux/prex/src/test/resources/pd/minimal_example.json
@@ -1,0 +1,25 @@
+{
+  "comment": "Note: VP, OIDC, DIDComm, or CHAPI outer wrapper would be here.",
+  "presentation_definition": {
+    "id": "32f54163-7166-48f1-93d8-ff217bdb0653",
+    "input_descriptors": [
+      {
+        "id": "wa_driver_license",
+        "name": "Washington State Business License",
+        "purpose": "We can only allow licensed Washington State business representatives into the WA Business Conference",
+        "constraints": {
+          "fields": [
+            {
+              "path": [
+                "$.credentialSubject.dateOfBirth",
+                "$.credentialSubject.dob",
+                "$.vc.credentialSubject.dateOfBirth",
+                "$.vc.credentialSubject.dob"
+              ]
+            }
+          ]
+        }
+      }
+    ]
+  }
+}
