+++ pollux/prex/src/test/resources/pd/single_group.json
@@ -1,0 +1,66 @@
+{
+  "comment": "VP, OIDC, DIDComm, or CHAPI outer wrapper here",
+  "presentation_definition": {
+    "id": "32f54163-7166-48f1-93d8-ff217bdb0653",
+    "submission_requirements": [{
+      "name": "Citizenship Information",
+      "rule": "pick",
+      "count": 1,
+      "from": "A"
+    }],
+    "input_descriptors": [
+      {
+        "id": "citizenship_input_1",
+        "name": "EU Driver's License",
+        "group": ["A"],
+        "constraints": {
+          "fields": [
+            {
+              "path": ["$.credentialSchema.id", "$.vc.credentialSchema.id"],
+              "filter": {
+                "type": "string",
+                "const": "https://eu.com/claims/DriversLicense.json"
+              }
+            },
+            {
+              "path": ["$.issuer", "$.vc.issuer", "$.iss"],
+              "purpose": "We can only accept digital driver's licenses issued by national authorities of member states or trusted notarial auditors.",
+              "filter": {
+                "type": "string",
+                "pattern": "^did:example:gov1$|^did:example:gov2$"
+              }
+            },
+            {
+              "path": ["$.credentialSubject.dob", "$.vc.credentialSubject.dob", "$.dob"],
+              "filter": {
+                "type": "string"
+              }
+            }
+          ]
+        }
+      },
+      {
+        "id": "citizenship_input_2",
+        "name": "US Passport",
+        "group": ["A"],
+        "constraints": {
+          "fields": [
+            {
+              "path": ["$.credentialSchema.id", "$.vc.credentialSchema.id"],
+              "filter": {
+                "type": "string",
+                "const": "hub://did:foo:123/Collections/schema.us.gov/passport.json"
+              }
+            },
+            {
+              "path": ["$.credentialSubject.birth_date", "$.vc.credentialSubject.birth_date", "$.birth_date"],
+              "filter": {
+                "type": "string"
+              }
+            }
+          ]
+        }
+      }
+    ]
+  }
+}
