+++ pollux/prex/src/test/resources/pd/two_filters_simplified.json
@@ -1,0 +1,43 @@
+{
+    "presentation_definition": {
+      "id": "Scalable trust example",
+      "input_descriptors": [
+        {
+          "id": "any type of credit card from any bank",
+          "name": "any type of credit card from any bank",
+          "purpose": "Please provide your credit card details",
+          "constraints": {
+            "fields": [
+              {
+                "path": [
+                  "$.termsOfUse.type"
+                ],
+                "filter": {
+                  "type": "string",
+                  "pattern": "^https://train.trust-scheme.de/info$"
+                }
+              },
+              {
+                "path": [
+                  "$.termsOfUse.trustScheme"
+                ],
+                "filter": {
+                  "type": "string",
+                  "pattern": "^worldbankfederation.com$"
+                }
+              },
+              {
+                "path": [
+                  "$.type"
+                ],
+                "filter": {
+                  "type": "string",
+                  "pattern": "^creditCard$"
+                }
+              }
+            ]
+          }
+        }
+      ]
+    }
+  }
